+++ src/Create.php
@@ -9,7 +9,12 @@ final class Create
     /**
      * Creates a promise for a value if the value is not a promise.
      *
-     * @param mixed $value Promise or value.
+     * @template TValue
+     * @template TPromise of PromiseInterface<mixed, mixed> = PromiseInterface<mixed, mixed>
+     *
+     * @param TValue|TPromise $value Promise or value.
+     *
+     * @return ($value is PromiseInterface ? TPromise : FulfilledPromise<TValue, mixed>)
      */
     public static function promiseFor($value): PromiseInterface
     {
@@ -34,7 +39,12 @@ final class Create
      * Creates a rejected promise for a reason if the reason is not a promise.
      * If the provided reason is a promise, then it is returned as-is.
      *
-     * @param mixed $reason Promise or reason.
+     * @template TReason
+     * @template TPromise of PromiseInterface<mixed, mixed> = PromiseInterface<mixed, mixed>
+     *
+     * @param TReason|TPromise $reason Promise or reason.
+     *
+     * @return ($reason is PromiseInterface ? TPromise : RejectedPromise<mixed, TReason>)
      */
     public static function rejectionFor($reason): PromiseInterface
     {
@@ -48,7 +58,9 @@ final class Create
     /**
      * Create an exception for a rejected promise value.
      *
-     * @param mixed $reason
+     * @template TReason
+     *
+     * @param TReason $reason
      */
     public static function exceptionFor($reason): \Throwable
     {
@@ -62,7 +74,12 @@ final class Create
     /**
      * Returns an iterator for the given value.
      *
-     * @param mixed $value
+     * @template TKey of array-key
+     * @template TValue
+     *
+     * @param iterable<TKey, TValue>|TValue $value
+     *
+     * @return \Iterator<TKey|int, TValue>
      */
     public static function iterFor($value): \Iterator
     {